docs: backport runnable code example fixes to v3 versioned docs
