+++ website/versioned_docs/version-3.4/03_guides/code/05_crawlee_playwright.py
@@ -54,7 +54,7 @@ async def main() -> None:
             proxy_configuration=proxy_configuration,
             request_handler=router,
             # Cap the crawl. Remove or increase the limit to follow all links.
-            max_requests_per_crawl=50,
+            max_requests_per_crawl=10,
             headless=True,
             browser_launch_options={'args': browser_args},
         )
