+++ website/versioned_docs/version-3.4/03_guides/code/07_scrapling_browser.py
@@ -79,7 +79,7 @@ async def main() -> None:
             await request_queue.add_request(Request.from_url(url))
 
         # Cap the crawl. Raise or remove the limit to follow more pages.
-        max_requests = 50
+        max_requests = 10
         handled_requests = 0
 
         # Open the browser once and reuse it for every page in the crawl.
