+++ website/versioned_docs/version-3.4/03_guides/code/08_crawl4ai.py
@@ -82,7 +82,7 @@ async def main() -> None:
             await request_queue.add_request(Request.from_url(url))
 
         # Cap the crawl; raise or remove to follow more pages.
-        max_requests = 50
+        max_requests = 10
         handled_requests = 0
 
         # Reuse one headless browser-backed crawler for every request.
